refactor: Separate core type definitions into types.ts
Moved shared data structure types (Subject, StoredNote, StoredQuiz, etc.) and re-exported AI flow types from `lib/actions.ts` to a new dedicated file `lib/types.ts`.

Updated `lib/actions.ts` and all relevant UI components (ExamsManager, NotesManager, QuizzesManager, subject page, analytics page, home page) to import these types from `lib/types.ts`.

This change improves separation of concerns and makes `lib/types.ts` the canonical source for shared type definitions, enhancing code organization and maintainability.
