+++ README.md
@@ -1,5 +1,110 @@
-# Firebase Studio
+# LearnFlow AI: Your Intelligent Learning Companion
 
-This is a NextJS starter in Firebase Studio.
+## Introduction
 
-To get started, take a look at src/app/page.tsx.
+LearnFlow AI is a next-generation learning platform designed to empower students, educators, and lifelong learners with AI-driven tools. Our mission is to make learning more efficient, engaging, and personalized. Whether you're struggling with complex concepts, preparing for exams, or simply curious to explore new topics, LearnFlow AI provides the resources you need to succeed.
+
+## Features
+
+LearnFlow AI offers a suite of powerful features, leveraging the capabilities of Google's Genkit AI:
+
+*   **Explain Terms:** Get clear and concise explanations for complex terms and concepts.
+*   **Extract Text from PDFs:** Easily extract text from PDF documents for further analysis or note-taking.
+*   **Find Relevant Articles:** Discover articles and resources relevant to your area of study.
+*   **Generate Dynamic Notes:** Create comprehensive and organized notes tailored to your learning needs.
+*   **Generate Exams and Analyze Results:** Generate practice exams and receive detailed analysis of your performance.
+*   **Generate Extra Readings:** Access supplementary reading materials to deepen your understanding of a topic.
+*   **Generate Quizzes:** Test your knowledge with AI-generated quizzes.
+
+## Technologies Used
+
+LearnFlow AI is built with a modern tech stack:
+
+*   **Next.js:** A React framework for building server-rendered and statically generated web applications.
+*   **Firebase:** A comprehensive platform for building web and mobile applications, used for backend services like authentication and database.
+*   **Genkit (Google AI):** Powers the AI-driven features of the platform.
+*   **Radix UI:** A collection of unstyled, accessible UI primitives.
+*   **Shadcn/ui:** Re-usable UI components built on Radix UI and Tailwind CSS.
+*   **Tailwind CSS:** A utility-first CSS framework for rapid UI development.
+
+## Getting Started
+
+### Prerequisites
+
+Before you begin, ensure you have the following installed:
+
+*   [Node.js](https://nodejs.org/) (version X.X.X or higher)
+*   [npm](https://www.npmjs.com/) or [yarn](https://yarnpkg.com/)
+*   Firebase account and project setup. Refer to the [Firebase documentation](https://firebase.google.com/docs/web/setup) for details.
+
+### Installation
+
+1.  **Clone the repository:**
+    ```bash
+    git clone https://github.com/your-username/learnflow-ai.git
+    cd learnflow-ai
+    ```
+2.  **Install dependencies:**
+    ```bash
+    npm install
+    # or
+    yarn install
+    ```
+3.  **Set up environment variables:**
+    Create a `.env.local` file in the root directory and add your Firebase configuration and other necessary API keys.
+    ```env
+    NEXT_PUBLIC_FIREBASE_API_KEY=your_api_key
+    NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=your_auth_domain
+    NEXT_PUBLIC_FIREBASE_PROJECT_ID=your_project_id
+    NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=your_storage_bucket
+    NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=your_messaging_sender_id
+    NEXT_PUBLIC_FIREBASE_APP_ID=your_app_id
+    # Add any Genkit related API keys if necessary
+    GENKIT_API_KEY=your_genkit_api_key
+    ```
+
+## Running the Application
+
+1.  **Start the Next.js development server:**
+    ```bash
+    npm run dev
+    # or
+    yarn dev
+    ```
+    Open [http://localhost:3000](http://localhost:3000) in your browser to see the application.
+
+2.  **Running the Genkit Development Server (if applicable):**
+    If your project uses Genkit with a local development server for AI flows, you might need to run it separately.
+    ```bash
+    # Command to start Genkit development server (e.g., genkit start)
+    # Please refer to your project's specific Genkit setup.
+    ```
+    Ensure your Next.js application is configured to communicate with the Genkit server endpoint.
+
+## Project Structure (Optional)
+
+A brief overview of the main directories:
+
+*   `app/`: Contains the core application logic, routes, and UI components (using Next.js App Router).
+*   `components/`: Shared UI components used across the application.
+*   `lib/`: Utility functions, Firebase configuration, and Genkit integrations.
+*   `flows/`: (If using Genkit) Definitions of AI flows.
+*   `public/`: Static assets like images and fonts.
+*   `styles/`: Global styles and Tailwind CSS configuration.
+
+## Contributing (Optional)
+
+We welcome contributions to LearnFlow AI! If you'd like to contribute, please follow these steps:
+
+1.  Fork the repository.
+2.  Create a new branch (`git checkout -b feature/your-feature-name`).
+3.  Make your changes and commit them (`git commit -m 'Add some feature'`).
+4.  Push to the branch (`git push origin feature/your-feature-name`).
+5.  Open a Pull Request.
+
+Please ensure your code adheres to our coding standards and includes tests where applicable.
+
+## License (Optional)
+
+This project is licensed under the MIT License. See the `LICENSE` file for details.
+(Note: You would need to create a `LICENSE` file with the MIT License text if you choose this license).
