+++ src/app/subjects/[subjectId]/_components/QuizzesManager.tsx
@@ -13,7 +13,8 @@ import { Label } from '@/components/ui/label';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter, DialogDescription } from '@/components/ui/dialog';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
 import { useToast } from '@/hooks/use-toast';
-import { generateQuizAction, explainTermAction, type Flashcard, type ExplainTermInput, type ExplainTermOutput, type StoredQuiz } from '@/lib/actions';
+import { generateQuizAction, explainTermAction } from '@/lib/actions';
+import type { Flashcard, ExplainTermInput, ExplainTermOutput, StoredQuiz } from '@/lib/types';
 import FileUpload from '@/components/common/FileUpload';
 import { HelpCircle, Sparkles, Tags, CheckCircle, AlertTriangle, RotateCcw, BookOpenText, ExternalLink, Trash2, Eye, Edit, Save } from 'lucide-react';
 import { Badge } from '@/components/ui/badge';
